toai

+++ muabansach/lib/model/Giohang.dart
@@ -11,9 +11,9 @@ class GioHang {
 
   factory GioHang.fromJson(Map<String, dynamic> json) {
     return GioHang(
-      id: json['id'] as int,
-      sachId: json['id_sach'] as int, // Sử dụng id_sach để lấy thông tin của sách
-      idNguoiDung: json['id_nguoidung'] as int,
+      id: json['id'],
+      sachId: json['id_sach'] ,
+      idNguoiDung: json['id_nguoidung'],
     );
   }
 }
